+++ flash/core/integrations/icevision/transforms.py
@@ -14,6 +14,7 @@
 from dataclasses import dataclass
 from typing import Any, Callable, Dict, List, Tuple
 
+import numpy as np
 from torch import nn
 
 from flash.core.data.io.input import DataKeys
@@ -24,7 +25,7 @@ if _ICEVISION_AVAILABLE:
     from icevision.core import tasks
     from icevision.core.bbox import BBox
     from icevision.core.keypoints import KeyPoints
-    from icevision.core.mask import EncodedRLEs, MaskArray
+    from icevision.core.mask import Mask, MaskArray
     from icevision.core.record import BaseRecord
     from icevision.core.record_components import (
         BBoxesRecordComponent,
@@ -37,12 +38,18 @@ if _ICEVISION_AVAILABLE:
     )
     from icevision.data.prediction import Prediction
     from icevision.tfms import A
+else:
+    MaskArray = object
 
 if _ICEVISION_AVAILABLE and _ICEVISION_GREATER_EQUAL_0_11_0:
-    from icevision.core.mask import MaskFile
     from icevision.core.record_components import InstanceMasksRecordComponent
 elif _ICEVISION_AVAILABLE:
     from icevision.core.record_components import MasksRecordComponent
+
+
+def _split_mask_array(mask_array: MaskArray) -> List[MaskArray]:
+    """Utility to split a single ``MaskArray`` object into a list of ``MaskArray`` objects (one per mask)."""
+    return [MaskArray(mask) for mask in mask_array.data]
 
 
 def to_icevision_record(sample: Dict[str, Any]):
@@ -57,6 +64,19 @@ def to_icevision_record(sample: Dict[str, Any]):
     component = ClassMapRecordComponent(tasks.detection)
     component.set_class_map(metadata.get("class_map", None))
     record.add_component(component)
+
+    if isinstance(sample[DataKeys.INPUT], str):
+        input_component = FilepathRecordComponent()
+        input_component.set_filepath(sample[DataKeys.INPUT])
+    else:
+        if "filepath" in metadata:
+            input_component = FilepathRecordComponent()
+            input_component.filepath = metadata["filepath"]
+        else:
+            input_component = ImageRecordComponent()
+        input_component.composite = record
+        input_component.set_img(sample[DataKeys.INPUT])
+    record.add_component(input_component)
 
     if "labels" in sample[DataKeys.TARGET]:
         labels_component = InstancesLabelsRecordComponent()
@@ -73,19 +93,19 @@ def to_icevision_record(sample: Dict[str, Any]):
         record.add_component(component)
 
     if _ICEVISION_GREATER_EQUAL_0_11_0:
-        mask_array = sample[DataKeys.TARGET].get("mask_array", None)
+        # mask_array = sample[DataKeys.TARGET].get("mask_array", None)
         masks = sample[DataKeys.TARGET].get("masks", None)
 
-        if mask_array is not None or masks is not None:
+        if masks is not None:
             component = InstanceMasksRecordComponent()
 
-            if masks is not None:
-                masks = [MaskFile(mask) for mask in masks]
-                component.set_masks(masks)
-
-            if mask_array is not None:
-                mask_array = MaskArray(mask_array)
-                component.set_mask_array(mask_array)
+            if masks is not None and len(masks) > 0:
+                if isinstance(masks[0], Mask):
+                    component.set_masks(masks)
+                else:
+                    mask_array = MaskArray(np.stack(masks, axis=0))
+                    component.set_mask_array(mask_array)
+                    component.set_masks(_split_mask_array(mask_array))
 
             record.add_component(component)
     else:
@@ -110,19 +130,6 @@ def to_icevision_record(sample: Dict[str, Any]):
         component.set_keypoints(keypoints)
         record.add_component(component)
 
-    if isinstance(sample[DataKeys.INPUT], str):
-        input_component = FilepathRecordComponent()
-        input_component.set_filepath(sample[DataKeys.INPUT])
-    else:
-        if "filepath" in metadata:
-            input_component = FilepathRecordComponent()
-            input_component.filepath = metadata["filepath"]
-        else:
-            input_component = ImageRecordComponent()
-        input_component.composite = record
-        input_component.set_img(sample[DataKeys.INPUT])
-    record.add_component(input_component)
-
     return record
 
 
@@ -142,26 +149,15 @@ def from_icevision_detection(record: "BaseRecord"):
             for bbox in detection.bboxes
         ]
 
-    mask_array = (
-        getattr(detection, "mask_array", None) if _ICEVISION_GREATER_EQUAL_0_11_0 else getattr(detection, "masks", None)
-    )
-    if mask_array is not None:
-        if isinstance(mask_array, EncodedRLEs):
-            mask_array = mask_array.to_mask(record.height, record.width)
-
-        if isinstance(mask_array, MaskArray):
-            result["mask_array"] = mask_array.data
-        else:
-            raise RuntimeError("Mask arrays are expected to be a MaskArray or EncodedRLEs.")
-
     masks = getattr(detection, "masks", None)
-    if masks is not None and _ICEVISION_GREATER_EQUAL_0_11_0:
-        result["masks"] = []
-        for mask in masks:
-            if isinstance(mask, MaskFile):
-                result["masks"].append(mask.filepath)
-            else:
-                raise RuntimeError("Masks are expected to be MaskFile objects.")
+    mask_array = getattr(detection, "mask_array", None)
+    if mask_array is not None or not _ICEVISION_GREATER_EQUAL_0_11_0:
+        if not isinstance(mask_array, MaskArray) or len(mask_array.data) == 0:
+            mask_array = MaskArray.from_masks(masks, record.height, record.width)
+
+        result["masks"] = [mask.data[0] for mask in _split_mask_array(mask_array)]
+    elif masks is not None:
+        result["masks"] = masks  # Note - this doesn't unpack IceVision objects
 
     if hasattr(detection, "keypoints"):
         keypoints = detection.keypoints